Объединила модули по отрисовке полноэкранного изображения в один fullsize-photo
