+++ js/fullsize-photo.js
@@ -1,4 +1,5 @@
 import {isEscKeydown} from './utils.js';
+
 const IMG_WIDTH = 35;
 const IMG_HEIGHT = 35;
 
@@ -7,7 +8,10 @@ const body = document.querySelector('body');
 const arrayComments = fullsizePhoto.querySelector('.social__comments');
 const countComments = fullsizePhoto.querySelector('.current-comments-count');
 const loadingComments = fullsizePhoto.querySelector('.comments-loader');
+const close = fullsizePhoto.querySelector('#picture-cancel');
 
+
+// Функции по подстановке в DOM-элемены значений/данных, взятых из массива
 
 // Функция по отрисовке/подстановке одного комментария
 const drawComment = (object) => {
@@ -107,4 +111,55 @@ const drawFullsizePhoto = (photo) => {
   body.classList.add('.modal-open');
 };
 
-export {fullsizePhoto, body, arrayComments, countComments, loadingComments, drawFullsizePhoto};
+
+//Функция для отображения миниатюры, по которой кликнули, в полноэкранный вид
+
+const showFullsizePhoto = (data) => {
+  // Поиск массива в DOM, чтобы его перебрать и навесить обработчик клика
+  const photos = document.querySelectorAll('.picture');
+
+  photos.forEach((photo) => {
+    photo.addEventListener('click', (evt) => {
+      drawFullsizePhoto(data[evt.currentTarget.dataset.id]); // photo должно браться из массива photos, который мы находим на странице, после отрисовки данных с сервера
+      toCreateCloseFullsizePhotoEvventListener(); // Создаю обработчики полноэкранного фото
+    });
+  });
+};
+
+
+// Закрываю окно полноэкранного отображения картинки
+
+const closePhoto = () => {
+  body.classList.remove('.modal-open');
+  fullsizePhoto.classList.add('hidden');
+  arrayComments.textContent = '';
+  countComments.textContent = '0';
+  loadingComments.classList.remove('hidden');
+  toDeleteCloseFullsizePhotoEvventListener(); // Удаляю обработчики закрытия полноэкранного изображения
+  document.onkeydown = null;
+};
+
+const toCloseFullsizePhoto = (evt) => {
+  evt.preventDefault();
+  closePhoto();
+};
+
+const toEscFullScreen = (evt) => {
+  if (isEscKeydown) {
+    evt.preventDefault();
+    closePhoto();
+  }
+};
+
+function toCreateCloseFullsizePhotoEvventListener() { // Создаю обработчики закрытия полноэкранного изображения
+  close.addEventListener('click', toCloseFullsizePhoto);
+  document.addEventListener('keydown', toEscFullScreen);
+}
+
+function toDeleteCloseFullsizePhotoEvventListener() { // Удаляю обработчики закрытия полноэкранного изображения
+  close.removeEventListener('click', toCloseFullsizePhoto);
+  document.removeEventListener('keydown', toEscFullScreen);
+}
+
+// export {fullsizePhoto, body, arrayComments, countComments, loadingComments, showFullsizePhoto};
+export {showFullsizePhoto};
