+++ js/remove-event-listeners.js
@@ -1,8 +1,9 @@
 import {controlSmaller, toReducePhoto, controlBigger, toEnlargePhoto} from './form-change-size-img.js';
 import {toResetEffects, effects, selectedEffect} from './form-effects-img.js';
 import {form, toSubmitForm} from './form-validation.js';
-import {body} from './fullsize-photo.js';
-import {buttonPopupSuccess, removePopupSuccess, removeToEscPopupSuccess, buttonPopupError, removePopupError, removeToEscPopupError, buttonPopupErrorLoad, removePopupErrorLoad, removeToEscPopupErrorLoad} from './popup-close.js';
+// import {close, toCloseFullsizePhoto, toEscFullScreen} from './fullsize-photo-close.js';
+// import {body} from './fullsize-photo.js';
+// import {buttonPopupSuccess, removePopupSuccess, removeToEscPopupSuccess, buttonPopupError, removePopupError, removeToEscPopupError, buttonPopupErrorLoad, removePopupErrorLoad, removeToEscPopupErrorLoad} from './popup-close.js';
 
 const toDeleteCloseFormEventListeners = () => {
   controlSmaller.removeEventListener('click', toReducePhoto);
@@ -12,22 +13,28 @@ const toDeleteCloseFormEventListeners = () => {
   toResetEffects();
 };
 
-const toDeleteClosePopupSuccessEventListeners = () => {
-  buttonPopupSuccess.removeEventListener('click', removePopupSuccess());
-  document.removeEventListener('keydown', removeToEscPopupSuccess());
-  body.removeEventListener('click', removePopupSuccess());
+const toDeleteCloseFullsizePhotoEvventListener = () => {
+  close.removeEventListener('click', toCloseFullsizePhoto);
+  document.removeEventListener('keydown', toEscFullScreen);
 };
 
-const toDeleteClosePopupErrorEventListeners = () => {
-  buttonPopupError.removeEventListener('click', removePopupError());
-  document.removeEventListener('keydown', removeToEscPopupError());
-  body.removeEventListener('click', removePopupError());
-};
+// const toDeleteClosePopupSuccessEventListeners = () => {
+//   buttonPopupSuccess.removeEventListener('click', removePopupSuccess());
+//   document.removeEventListener('keydown', removeToEscPopupSuccess());
+//   body.removeEventListener('click', removePopupSuccess());
+// };
 
-const toDeleteClosePopupErrorLoadEventListeners = () => {
-  buttonPopupErrorLoad.addEventListener('click', removePopupErrorLoad());
-  document.addEventListener('keydown', removeToEscPopupErrorLoad());
-  body.addEventListener('click', removePopupErrorLoad());
-};
+// const toDeleteClosePopupErrorEventListeners = () => {
+//   buttonPopupError.removeEventListener('click', removePopupError());
+//   document.removeEventListener('keydown', removeToEscPopupError());
+//   body.removeEventListener('click', removePopupError());
+// };
 
-export {toDeleteCloseFormEventListeners, toDeleteClosePopupSuccessEventListeners, toDeleteClosePopupErrorEventListeners, toDeleteClosePopupErrorLoadEventListeners};
+// const toDeleteClosePopupErrorLoadEventListeners = () => {
+//   buttonPopupErrorLoad.addEventListener('click', removePopupErrorLoad());
+//   document.addEventListener('keydown', removeToEscPopupErrorLoad());
+//   body.addEventListener('click', removePopupErrorLoad());
+// };
+
+// export {toDeleteCloseFormEventListeners, toDeleteClosePopupSuccessEventListeners, toDeleteClosePopupErrorEventListeners, toDeleteClosePopupErrorLoadEventListeners};
+export {toDeleteCloseFullsizePhotoEvventListener, toDeleteCloseFormEventListeners};
